Move Annotation resolvers into TypeScript. Additional changes: * In colocalizationCoeff, database will now fall back to a default value * In colocalizationCoeff, ``'getColocalizationCoeff' in hit` is used     instead of `hit.getColocalizationCoeff != null` so that TypeScript     can use it to infer the ESAnnotationWithColoc subclass. * Fixed definition of config.img_upload.categories
