+++ metaspace/graphql/src/modules/annotation/controller/index.ts
@@ -1,0 +1,7 @@
+import Annotation from './Annotation';
+import {IResolvers} from 'graphql-tools';
+import {Context} from '../../../context';
+
+export const Resolvers = {
+  Annotation,
+} as IResolvers<any, Context>;
